Refactor test structure for query parameters
