+++ tests/test_query/__init__.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+import httpx
+
+
+@dataclass(frozen=True)
+class QueryParamCase:
+    expected: httpx.QueryParams
